requirements.txt
discord.py
aiosqlite
discord-py-slash-command